Kill http server process upon electron exit
When all windows are closed we want the electron app to end the rust http server.
The current idea is to have the rust server write its PID to a file.
Then the Electron UI can read it and on close hit the server with a SIGINT.
Still in progress...